# check for private keys in CI and speed up checks