--- a/.github/workflows/build.yml
+++ b/.github/workflows/build.yml
@@ -51,6 +51,18 @@ jobs:
         run: |
           mypy .
 
+  pre-commit:
+    runs-on: ubuntu-latest
+    env:
+      SKIP: "ruff"
+    steps:
+      - uses: actions/checkout@v4
+      - uses: actions/setup-python@v5
+        with:
+          python-version: "3.11"
+          cache: pip
+      - uses: pre-commit/action@v3.0.1
+
   test:
     runs-on: ubuntu-latest
     strategy: