--- a/.pre-commit-config.yaml
+++ b/.pre-commit-config.yaml
@@ -14,6 +14,8 @@ repos:
     hooks:
       - id: check-added-large-files
       - id: check-json
+        # The following are JSONC files.
+        exclude: '(^\.vscode/|^tools/vscode/|jsconfig\.json$|tsconfig\.json$)'
       - id: check-yaml
       - id: debug-statements
       - id: detect-private-key