--- a/docs/_max_samples.md
+++ b/docs/_max_samples.md
@@ -4,4 +4,3 @@ Another consideration is `max_samples`, which is the maximum number of samples t
 By default, Inspect sets the value of `max_samples` to `max_connections + 1` (note that it would rarely make sense to set it _lower_ than `max_connections`). The default `max_connections` is 10, which will typically result in samples being written to the log frequently. On the other hand, setting a very large `max_connections` (e.g. 100 `max_connections` for a dataset with 100 samples) may result in very few recoverable samples in the case of an interruption. 
 
 {{< include _setting_max_samples.md >}}
-