--- a/docs/_store_typing.md
+++ b/docs/_store_typing.md
@@ -49,4 +49,3 @@ from inspect_ai.util import Store
 store = Store()
 activity = Activity(store=store)
 ```
-