--- a/docs/evals/index.qmd
+++ b/docs/evals/index.qmd
@@ -22,4 +22,3 @@ These evals can be `pip` installed and run with a single command against any mod
 
 ::: {#evals}
 :::
-