--- a/docs/models.qmd
+++ b/docs/models.qmd
@@ -122,4 +122,3 @@ async with get_model("openai/gpt-4o") as model:
 
 - [Reasoning](reasoning.qmd) documents the additional options and data available for reasoning models.
  
-