--- a/docs/reference/inspect_ai.approval.qmd
+++ b/docs/reference/inspect_ai.approval.qmd
@@ -17,4 +17,3 @@ title: "inspect_ai.approval"
 ## Decorator
 
 ### approver
-