--- a/docs/reference/inspect_ai.dataset.qmd
+++ b/docs/reference/inspect_ai.dataset.qmd
@@ -15,4 +15,3 @@ title: "inspect_ai.dataset"
 ### RecordToSample
 ### Dataset
 ### MemoryDataset
-