--- a/docs/reference/inspect_ai.log.qmd
+++ b/docs/reference/inspect_ai.log.qmd
@@ -56,12 +56,3 @@ title: "inspect_ai.log"
 ### InfoEvent
 ### StepEvent
 ### SubtaskEvent
-
-
-
-
-
-
-
-
-