--- a/docs/reference/inspect_ai.model.qmd
+++ b/docs/reference/inspect_ai.model.qmd
@@ -49,6 +49,3 @@ title: "inspect_ai.model"
 
 ### modelapi
 ### ModelAPI
-
-
-