--- a/docs/reference/inspect_ai.scorer.qmd
+++ b/docs/reference/inspect_ai.scorer.qmd
@@ -48,5 +48,3 @@ title: "inspect_ai.scorer"
 ### scorer
 ### metric
 ### score_reducer
-
-