--- a/docs/reference/inspect_ai.tool.qmd
+++ b/docs/reference/inspect_ai.tool.qmd
@@ -31,4 +31,3 @@ title: "inspect_ai.tool"
 ## Decorator
 
 ### tool
-