--- a/docs/reference/inspect_ai.util.qmd
+++ b/docs/reference/inspect_ai.util.qmd
@@ -41,4 +41,3 @@ title: "inspect_ai.util"
 ### SandboxEnvironment
 ### SandboxConnection
 ### sandboxenv
-