--- a/docs/reference/inspect_cache.qmd
+++ b/docs/reference/inspect_cache.qmd
@@ -1,10 +1,3 @@
 ---
 title: inspect cache
 ---
-
-
-
-
-
-
-