--- a/docs/reference/inspect_eval-retry.qmd
+++ b/docs/reference/inspect_eval-retry.qmd
@@ -1,6 +1,3 @@
 ---
 title: inspect eval-retry
 ---
-
-
-