--- a/docs/reference/inspect_eval-set.qmd
+++ b/docs/reference/inspect_eval-set.qmd
@@ -1,6 +1,3 @@
 ---
 title: inspect eval-set
 ---
-
-
-