--- a/docs/reference/inspect_eval.qmd
+++ b/docs/reference/inspect_eval.qmd
@@ -1,4 +1,3 @@
 ---
 title: inspect eval
 ---
-