--- a/docs/reference/inspect_info.qmd
+++ b/docs/reference/inspect_info.qmd
@@ -1,12 +1,3 @@
 ---
 title: inspect info
 ---
-
-
-
-
-
-
-
-
-