--- a/docs/reference/inspect_list.qmd
+++ b/docs/reference/inspect_list.qmd
@@ -1,10 +1,3 @@
 ---
 title: inspect list
 ---
-
-
-
-
-
-
-