--- a/docs/reference/inspect_log.qmd
+++ b/docs/reference/inspect_log.qmd
@@ -1,7 +1,3 @@
 ---
 title: inspect log
 ---
-
-
-
-