--- a/docs/reference/inspect_sandbox.qmd
+++ b/docs/reference/inspect_sandbox.qmd
@@ -1,9 +1,3 @@
 ---
 title: inspect sandbox
 ---
-
-
-
-
-
-