--- a/docs/reference/inspect_score.qmd
+++ b/docs/reference/inspect_score.qmd
@@ -1,6 +1,3 @@
 ---
 title: inspect score
 ---
-
-
-