--- a/docs/reference/inspect_trace.qmd
+++ b/docs/reference/inspect_trace.qmd
@@ -1,9 +1,3 @@
 ---
 title: inspect trace
 ---
-
-
-
-
-
-