--- a/docs/reference/inspect_view.qmd
+++ b/docs/reference/inspect_view.qmd
@@ -1,4 +1,3 @@
 ---
 title: "inspect view"
 ---
-