--- a/docs/theme.scss
+++ b/docs/theme.scss
@@ -138,4 +138,3 @@ img.navbar-logo {
 .ref-definition {
     font-weight: normal;
 }
-