--- a/src/inspect_ai/tool/_tools/_computer/_resources/README.md
+++ b/src/inspect_ai/tool/_tools/_computer/_resources/README.md
@@ -27,4 +27,3 @@ The image extends this minimal functionality by adding a few basic applications 
 ## Desktop Directory
 
 The `Desktop` directory contains launchers for VS Code, Firefox and XPaint.
-