--- a/src/inspect_ai/tool/_tools/_computer/_resources/entrypoint/xfce_startup.sh
+++ b/src/inspect_ai/tool/_tools/_computer/_resources/entrypoint/xfce_startup.sh
@@ -10,4 +10,3 @@ done
 
 echo "XFCE4 is fully started!"
 touch /tmp/xfce_started
-