--- a/tests/util/sandbox/sandbox_setup_error.sh
+++ b/tests/util/sandbox/sandbox_setup_error.sh
@@ -1,4 +1,3 @@
 #!/usr/bin/env bash
 
 exit 1
-