--- a/tools/vscode/src/@types/hooks.d.ts
+++ b/tools/vscode/src/@types/hooks.d.ts
@@ -73,5 +73,3 @@ declare module 'positron' {
     readonly previewPanel: PreviewPanel;
   }
 }
-
-