--- a/tools/vscode/src/core/appdirs.ts
+++ b/tools/vscode/src/core/appdirs.ts
@@ -138,5 +138,3 @@ function windowsUserCacheDir(appName: string) {
   const dir = process.env.LOCALAPPDATA || "";
   return path.join(dir, "Temp", appName, appName);
 }
-
-