--- a/tools/vscode/src/core/git.ts
+++ b/tools/vscode/src/core/git.ts
@@ -57,4 +57,3 @@ function writeGitignore(dir: string, lines: string[]) {
     { encoding: "utf-8" }
   );
 }
-