--- a/tools/vscode/src/core/jsonrpc.ts
+++ b/tools/vscode/src/core/jsonrpc.ts
@@ -241,5 +241,3 @@ function methodNotFoundResponse(request: JsonRpcRequest) {
     `Method '${request.method}' not found.`
   );
 }
-
-