--- a/tools/vscode/src/core/python/code.ts
+++ b/tools/vscode/src/core/python/code.ts
@@ -48,5 +48,3 @@ const kReservedWords = [
   "with",
   "yield",
 ];
-
-