--- a/tools/vscode/src/extension.ts
+++ b/tools/vscode/src/extension.ts
@@ -203,5 +203,3 @@ const checkInspectVersion = async () => {
     }
   }
 };
-
-