--- a/tools/vscode/src/inspect/index.ts
+++ b/tools/vscode/src/inspect/index.ts
@@ -1,5 +1,3 @@
 
 
 export { initInspectProps, inspectVersionDescriptor as inspectVersion } from './props';
-
-