--- a/tools/vscode/src/providers/activity-bar/activity-bar-provider.ts
+++ b/tools/vscode/src/providers/activity-bar/activity-bar-provider.ts
@@ -56,4 +56,3 @@ export async function activateActivityBar(
 
   return [...outlineCommands, ...taskConfigCommands, ...logsCommands];
 }
-