--- a/tools/vscode/src/providers/activity-bar/log-listing/log-listing-data.ts
+++ b/tools/vscode/src/providers/activity-bar/log-listing/log-listing-data.ts
@@ -201,4 +201,3 @@ function formatPrettyDateTime(date: Date) {
   // For other years, include the year
   return format(date, "MMM d yyyy, h:mmaaa");
 }
-