--- a/tools/vscode/src/providers/activity-bar/log-listing/log-listing-mru.ts
+++ b/tools/vscode/src/providers/activity-bar/log-listing/log-listing-mru.ts
@@ -44,5 +44,3 @@ export class LogListingMRU {
     return this.context_.workspaceState.get<string[]>(kMruKey) || [];
   }
 }
-
-