--- a/tools/vscode/src/providers/activity-bar/webview/env-config-webview.css
+++ b/tools/vscode/src/providers/activity-bar/webview/env-config-webview.css
@@ -61,4 +61,3 @@
   height: 4px;
   line-height: 4px;
 }
-