--- a/tools/vscode/src/providers/activity-bar/webview/webview-utils.ts
+++ b/tools/vscode/src/providers/activity-bar/webview/webview-utils.ts
@@ -67,6 +67,3 @@ export function setControlsVisible(id: string, visible: boolean) {
     controls?.classList.add("hidden");
   }
 }
-
-
-