--- a/tools/vscode/src/providers/inspect/inspect-eval-commands.ts
+++ b/tools/vscode/src/providers/inspect/inspect-eval-commands.ts
@@ -30,4 +30,3 @@ export class DebugEvalCommand implements Command {
   private static readonly id = "inspect.debugTask";
   public readonly id = DebugEvalCommand.id;
 }
-