--- a/tools/vscode/src/providers/logview/commands.ts
+++ b/tools/vscode/src/providers/logview/commands.ts
@@ -45,4 +45,3 @@ class ShowLogviewCommand implements Command {
   private static readonly id = "inspect.inspectView";
   public readonly id = ShowLogviewCommand.id;
 }
-