--- a/tools/vscode/src/providers/logview/logview-panel.ts
+++ b/tools/vscode/src/providers/logview/logview-panel.ts
@@ -248,4 +248,3 @@ function webviewPanelJsonRpcServer(
   };
   return jsonRpcPostMessageServer(target, methods);
 }
-