--- a/tools/vscode/src/providers/logview/logview-view.ts
+++ b/tools/vscode/src/providers/logview/logview-view.ts
@@ -358,4 +358,3 @@ class InspectViewWebview extends InspectWebview<LogviewState> {
     return this.logviewPanel_.getHtml(state);
   }
 }
-