--- a/tools/vscode/src/providers/workspace/workspace-state-provider.ts
+++ b/tools/vscode/src/providers/workspace/workspace-state-provider.ts
@@ -75,4 +75,3 @@ function taskKey(file: string, task?: string) {
 function modelKey(provider: string) {
   return `provider-${provider}`;
 }
-