--- a/tools/vscode/tools/ts-to-mjs/src/jsonrpc.ts
+++ b/tools/vscode/tools/ts-to-mjs/src/jsonrpc.ts
@@ -235,5 +235,3 @@ function methodNotFoundResponse(request: JsonRpcRequest) {
     `Method '${request.method}' not found.`
   );
 }
-
-