… .pylintrc
… .vscode/extensions.json
… .vscode/settings.json
… docs/.gitignore
… docs/_container_limits.md
… docs/_quarto.yml
… docs/_sandboxenv-interface.md
… docs/_tools-scaffold.md
… docs/agents-api.qmd
… docs/approval.qmd
… docs/caching.qmd
… docs/errors-and-limits.qmd
… docs/eval-logs.qmd
… docs/eval-sets.qmd
… docs/evals/evals.css
… docs/human-agent.qmd
… docs/index.qmd
… docs/log-viewer.qmd
… docs/multimodal.qmd
… docs/options.qmd
… docs/providers.qmd
… docs/reasoning.qmd
… docs/reference/_sidebar.yml
… docs/reference/filter/commands.py
… docs/scorers.qmd
… docs/solvers.qmd
… docs/tasks.qmd
… docs/tracing.qmd
… docs/vscode.qmd
… examples/bridge/autogen/dataset.json