I'm looking into running this project locally and noticed some quick easy wins.
- There are some automated checks that don't seem to run in CI eg private key check.
- Some files that were not following what was laid out by those (pre-commit) checks
- Running checks could be easily made faster in some scenarios by allowing concurrent runs (~2x). (`make -j`)

## This PR contains:
- [x] Changes to dev-tools e.g. CI config / github tooling

### What is the current behavior? (You can also link to an open issue here)
- pre-commit hooks are _not_ enforced in CI

### What is the new behavior?
- pre-commit hooks are enforced in CI. If people don't agree on what some of those hooks are doing there is
a knob to turn some of them off.

### Other information:

I'd suggest reviewing commit by commit rather than starting with the files view as the majority of the changes are automated and isolated in one of the commits.

```bash
(inspect_ai) hbp14 ➜ c/inspect_ai git:(z-init-buildtest) gco main; g rev-parse HEAD; rm -rf .mypy_cache; time make check -j > /dev/null 
Switched to branch 'main'
Your branch is up to date with 'origin/main'.
2db59524a6a3d06c52592cf08f9c349da3984265
make check -j > /dev/null  28.13s user 1.87s system 97% cpu 30.909 total
Time: 0h:00m:31s

(inspect_ai) hbp14 ➜ c/inspect_ai git:(main) gco z-init-buildtest ; g rev-parse HEAD; rm -rf .mypy_cache; time make check -j > /dev/null    
Switched to branch 'z-init-buildtest'
25db832db333038679a74dbd2b446ad9f78a5326
make check -j > /dev/null  30.52s user 2.18s system 181% cpu 18.014 total
Time: 0h:00m:18s
```