@jjallaire-aisi 2025-02-19
Thanks for this! We just merged a variation of it without the end of line hook: https://github.com/UKGovernmentBEIS/inspect_ai/pull/1351